+++ com.unity.cinemachine/Runtime/Debug/CinemachineDebug.cs
@@ -39,9 +39,10 @@ namespace Unity.Cinemachine
         
 #if UNITY_EDITOR
         const string k_DebugUIName = "CinemachineDebugUI";
+        static Dictionary<Camera, VisualElement> s_CameraViewRectContainers = new();
         static GameObject s_UIDocumentHolder;
         static UIDocument s_UIDocument;
-        public static VisualElement CreateRuntimeUIContainer()
+        public static VisualElement CreateRuntimeUIContainer(Camera outputCamera)
         {
             if (s_UIDocumentHolder == null)
             {
@@ -51,7 +52,7 @@ namespace Unity.Cinemachine
                     s_UIDocumentHolder = new GameObject(k_DebugUIName)
                     {
                         hideFlags = HideFlags.NotEditable | HideFlags.HideInHierarchy,
-                        tag = "EditorOnly"
+                        tag = "EditorOnly" // TODO: we may want runtime too
                     };
                 }
             }
@@ -62,33 +63,37 @@ namespace Unity.Cinemachine
                 const string path = "Packages/com.unity.cinemachine/Runtime/UI/";
                 s_UIDocument.panelSettings = AssetDatabase.LoadAssetAtPath<PanelSettings>(path + "CinemachinePanelSettings.asset");
             }
-            
+
+            if (s_CameraViewRectContainers.ContainsKey(outputCamera))
+                return s_CameraViewRectContainers[outputCamera];
+
             var viewportContainer = new VisualElement
             {
                 name = "CinemachineDebugUI_ViewportContainer",
                 style = { position = new StyleEnum<Position>(Position.Absolute) }
             };
-            s_UIDocument.rootVisualElement?.Add(viewportContainer);
-            return viewportContainer;
-        }
-
-        /// <summary>Position the input visual element in the panel defined by s_ViewportContainer and the camera.</summary>
-        /// <param name="visualElement">The visual element to position. It must be a child of s_ViewportContainer</param>
-        /// <param name="camera">The camera that defines the view area.</param>
-        public static void PositionWithinCameraView(VisualElement visualElement, Camera camera)
-        {
-            if (s_UIDocument == null || s_UIDocument.rootVisualElement == null || camera == null)
-                return;
+            // need to use delayCall, because rootVisualElement may not be built at this point yet
+            EditorApplication.delayCall += () => s_UIDocument.rootVisualElement.Add(viewportContainer);
+            viewportContainer.schedule.Execute(() => PositionWithinCameraView(viewportContainer, outputCamera)).Every(0); // TODO: can we do better?
+            s_CameraViewRectContainers.Add(outputCamera, viewportContainer);
             
-            var panel = s_UIDocument.rootVisualElement.panel;
-            var screenPanelSpace = RuntimePanelUtils.ScreenToPanel(panel, new Vector2(Screen.width, Screen.height));
-            var viewport = camera.pixelRect;
-            var viewPortMaxPanelSpace = RuntimePanelUtils.ScreenToPanel(panel, viewport.max);
-            var viewPortMinPanelSpace = RuntimePanelUtils.ScreenToPanel(panel, viewport.min);
-            visualElement.style.top = new Length(screenPanelSpace.y - viewPortMaxPanelSpace.y, LengthUnit.Pixel);
-            visualElement.style.bottom = new Length(viewPortMinPanelSpace.y, LengthUnit.Pixel);
-            visualElement.style.left = new Length(viewPortMinPanelSpace.x, LengthUnit.Pixel);
-            visualElement.style.right = new Length(screenPanelSpace.x - viewPortMaxPanelSpace.x, LengthUnit.Pixel);
+            return viewportContainer;
+            
+            static void PositionWithinCameraView(VisualElement visualElement, Camera camera)
+            {
+                if (s_UIDocument == null || s_UIDocument.rootVisualElement == null || camera == null)
+                    return;
+            
+                var panel = s_UIDocument.rootVisualElement.panel;
+                var screenPanelSpace = RuntimePanelUtils.ScreenToPanel(panel, new Vector2(Screen.width, Screen.height));
+                var viewport = camera.pixelRect;
+                var viewPortMaxPanelSpace = RuntimePanelUtils.ScreenToPanel(panel, viewport.max);
+                var viewPortMinPanelSpace = RuntimePanelUtils.ScreenToPanel(panel, viewport.min);
+                visualElement.style.top = new Length(screenPanelSpace.y - viewPortMaxPanelSpace.y, LengthUnit.Pixel);
+                visualElement.style.bottom = new Length(viewPortMinPanelSpace.y, LengthUnit.Pixel);
+                visualElement.style.left = new Length(viewPortMinPanelSpace.x, LengthUnit.Pixel);
+                visualElement.style.right = new Length(screenPanelSpace.x - viewPortMaxPanelSpace.x, LengthUnit.Pixel);
+            }
         }
 #endif
 
@@ -125,78 +130,5 @@ namespace Unity.Cinemachine
                 s_AvailableStringBuilders = new List<StringBuilder>();
             s_AvailableStringBuilders.Add(sb);
         }
-        
-#if UNITY_EDITOR
-        internal class DebugText : IDisposable
-        {
-            Camera m_Camera;
-            VisualElement m_DebugUIContainer;
-            Label m_DebugLabel;
-            StyleColor m_OriginalTextColor;
-
-            public DebugText(Camera outputCamera)
-            {
-                m_Camera = outputCamera;
-                
-                m_DebugUIContainer = CreateRuntimeUIContainer();
-                m_DebugLabel = new Label
-                {
-                    style =
-                    {
-                        backgroundColor = new StyleColor(new Color(0.5f, 0.5f, 0.5f, 0.5f)),
-                        marginBottom = new StyleLength(new Length(2, LengthUnit.Pixel)),
-                        marginTop = new StyleLength(new Length(2, LengthUnit.Pixel)),
-                        marginLeft = new StyleLength(new Length(2, LengthUnit.Pixel)),
-                        marginRight = new StyleLength(new Length(2, LengthUnit.Pixel)),
-                        paddingBottom = new StyleLength(new Length(0, LengthUnit.Pixel)),
-                        paddingTop = new StyleLength(new Length(0, LengthUnit.Pixel)),
-                        paddingLeft = new StyleLength(new Length(0, LengthUnit.Pixel)),
-                        paddingRight = new StyleLength(new Length(0, LengthUnit.Pixel)),
-                        fontSize = new StyleLength(new Length(10, LengthUnit.Percent)),
-                        color = new StyleColor(Color.white),
-                        position = new StyleEnum<Position>(Position.Relative),
-                        alignSelf = new StyleEnum<Align>(Align.FlexStart)
-                    }
-                };
-                m_DebugUIContainer.Add(m_DebugLabel);
-                m_OriginalTextColor = m_DebugLabel.style.color;
-
-                m_DebugUIContainer.schedule.Execute(PositionElement).Every(0);
-            }
-
-            public void SetTextColor(Color color) => m_DebugLabel.style.color = new StyleColor(color);
-            public void RestoreOriginalTextColor() => m_DebugLabel.style.color = m_OriginalTextColor;
-            public void SetVisibility(bool v) => m_DebugUIContainer.visible = v;
-            public void SetText(string text) => m_DebugLabel.text = text;
-
-            public void Dispose()
-            {
-                ReleaseUnmanagedResources();
-                GC.SuppressFinalize(this);
-            }
-            
-            void PositionElement() => PositionWithinCameraView(m_DebugUIContainer, m_Camera);
-
-            ~DebugText()
-            {
-                ReleaseUnmanagedResources();
-            }
-            
-            void ReleaseUnmanagedResources()
-            {
-                if (m_DebugLabel != null)
-                {
-                    m_DebugLabel.RemoveFromHierarchy();
-                    m_DebugLabel = null;
-                }
-
-                if (m_DebugUIContainer != null)
-                {
-                    m_DebugUIContainer.RemoveFromHierarchy();
-                    m_DebugUIContainer = null;
-                }
-            }
-        }
-#endif
     }
 }
