+++ com.unity.cinemachine/Runtime/Debug/DebugText.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 01011f0717cf4b59b383b1fa591ddb07
+timeCreated: 1679931472
